bin/requirements.txt
requests==2.22.0
responses==0.10.7
pytest==5.3.1
snakebite-py3==3.0.5
pymongo==3.10.0
